Clean up redundant UI scripts
