--- Assets/Scripts/UI/DeactivatePlayerInputWhenEnabled.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12a269fcfc4d545479f62dab40d7fbea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
