--- Assets/Scripts/UI/HUD/PauseMenu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ffa46b99e3158470e9b1627a03758868
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
