--- Assets/Scripts/UI/HUD/PauseMenu/UIPauseMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 994c3543a40024d47a539f0e4f9c276b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
